+++ handlers/ticket-tailor-webhook/test/firstTest.test.ts
@@ -1,3 +1,5 @@
+import { handler } from "../src";
+
 /**
  * This is a unit test, it can be run by the `pnpm test` command, and will be run by the CI/CD pipeline
  *
@@ -6,5 +8,15 @@
  * The test is retained here simply to prevent the build from failing
  */
 test('my app', () => {
+
+	// {
+	// 	"buyer_details": {
+	// 	  "email": "john@example.com",
+	// 	  "first_name": "John",
+	// 	  "last_name": "Doe",
+	// 	  "name": "John Doe",
+	// 	  "phone": "07123456789"
+	// 	}
+	//   }
 	expect(1 + 1).toEqual(2);
 });
